% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densities.plot.R
\name{densities.plot}
\alias{densities.plot}
\title{Plot densities for multiple modes.}
\usage{
densities.plot(x, modes = 1:4, mark_modes = TRUE, in_one = TRUE *
  (length(modes) < 4))
}
\description{
Plot densities for multiple modes.
}
\examples{
densities.plot(c(rnorm(10), rnorm(10, mean = 10)))

}
