% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h.crit.R
\name{h.crit}
\alias{h.crit}
\title{Bandwidth calculation for density estimation}
\usage{
h.crit(x, k, prec = 6, density.fun = NULL)
}
\arguments{
\item{x}{vector of data}

\item{k}{number of modes}

\item{prec}{number of digits for precision of calculation}

\item{density.fun}{A function that returns a vector of density estimates}
}
\value{
the smallest value so that the gaussian kernel density estimate of the given data \code{x} has \code{k} modes.
}
\description{
Calculates the smallest value so that the gaussian kernel density estimate of the given data \code{x} has \code{k} modes.
The smaller you choose the bandwidth for a kernel density estimate, the larger the number of modes becomes. This function calculates the smallest value leading to a kernel density estimate with \code{k} number of modes.
}
