% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nr.modes.R
\name{nr.modes}
\alias{nr.modes}
\title{Number of modes}
\usage{
nr.modes(y)
}
\arguments{
\item{y}{vector of y-values of a density function}
}
\description{
Calculates the number of modes for given y-values of a density function.
}
