% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/silverman.plot.R
\name{silverman.plot}
\alias{silverman.plot}
\title{Plot p-values along different number of modes.}
\usage{
silverman.plot(x, kmin = 1, kmax = 5, alpha = 0.05, adjust = FALSE)
}
\arguments{
\item{x}{The vector of data to use for testing.}

\item{kmin, kmax}{The number of modes.}

\item{alpha}{Significance level, shown as a dashed line in the resulting plot.}

\item{adjust}{Perform p-value adjustment in case of \code{k=1} (see Hall and York).}
}
\description{
P-values are obtained using the \code{silverman.test} function with the number of modes ranging from \code{kmin} to \code{kmax}.
}
\examples{
silverman.plot(x = c(rpois(n = 50, lambda = 1), rnorm(n = 100, mean = 4)))

}
