% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/silverman.test.R
\name{silverman.test}
\alias{silverman.test}
\title{Silvermantest}
\usage{
silverman.test(x, k, R = 999, adjust = FALSE, digits = 6,
  density.fun = NULL)
}
\arguments{
\item{x}{vector of data}

\item{k}{number of modes for the null hypothesis}

\item{R}{number of bootstrap replications}

\item{adjust}{boolean to activate the adjusting of the p-value (valid if k=1) (see Hall and York)}

\item{digits}{number of digits of the p-value}

\item{density.fun}{A function that returns a vector of density estimates}
}
\value{
An object of the class Silvermantest (see: \code{\link{Silvermantest-class}}).
}
\description{
The silvermantest tests the null hypothesis that an underlying density has at most \code{k} modes.
}
\examples{
\dontrun{
x <- c(rnorm(50), rnorm(50, mean = 3, sd = 0.1))
silverman.test(x, k = 1)
}

}
