% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{Silvermantest-class}
\alias{Silvermantest-class}
\title{The Silvermantest-class.}
\description{
The Silvermantest-class.
}
\section{Slots}{

\describe{
\item{\code{data}}{The underlying test data}

\item{\code{p_value}}{The p-value of the test}

\item{\code{saved_seed}}{The random seed that was used during calculation}

\item{\code{k}}{The number of modes tested}
}}

